pyproject.toml
[project]
name = "nonebot-plugin-yoyogame"
version = "0.1.0"
description = "一个基于Nonebot开发的悠悠小游戏"
authors = [
    {name = "ChenXu233", email = "Woyerpa@outlook.com"},
]
dependencies = [
    "pandas>=2.2.2",
    "terminaltables>=3.1.10",
    "nonebot2>=2.3.3",
    "nonebot-adapter-onebot>=2.4.4",
]
requires-python = ">=3.10"

readme = "README.md"
license = {text = "GPL"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true
